Made upload profile picture in base64

+++ Api/Controllers/User/UserController.cs
@@ -262,7 +262,8 @@ public class UserController : ControllerBase
 
 
                 //Remove the current profile picture if exist
-                if (currentUser.ProfilePicturePath != null)
+                if (currentUser.ProfilePicturePath != null && currentUser.ProfilePicturePath !=
+                    "\\Upload\\ProfilePicture\\default_user_pic.png")
                 {
                     _pictureService.RemoveFile(currentUser.ProfilePicturePath);
                 }
@@ -286,6 +287,58 @@ public class UserController : ControllerBase
         }
 
         return Unauthorized("Failed");
+    }
+
+    [HttpPut]
+    [Authorize]
+    [Route("profilePicture/base64")]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+    public ActionResult<DtoOutputUser> UpdateProfilePictureBase64(DtoInputProfilePictureBase64 dto)
+    {
+        var userId = int.Parse(User.Identity?.Name);
+        var currentUser = _userService.FetchById(userId);
+        var filePath = "\\Upload\\ProfilePicture\\default_user_pic.png";
+
+
+        if (dto.ProfilePicture == null)
+        {
+            //Remove the current profile picture if exist
+            if (currentUser.ProfilePicturePath != null && currentUser.ProfilePicturePath !=
+                "\\Upload\\ProfilePicture\\default_user_pic.png")
+            {
+                _pictureService.RemoveFile(currentUser.ProfilePicturePath);
+            }
+        }
+        else
+        {
+            var basePath = "\\Upload\\ProfilePicture\\";
+
+            filePath = basePath + _pictureService.GenerateUniqueFileName(userId) +
+                       _pictureService.GetExtension(dto.ProfilePicture);
+
+            _pictureService.UploadBase64Picture(basePath, filePath, dto.ProfilePicture);
+        }
+
+
+        var dtoInputUpdateProfilePictureUser = new DtoInputUpdateProfilePictureUser
+        {
+            Id = userId,
+            ProfilePicturePath = filePath
+        };
+
+        var oldProfilePicturePath = currentUser.ProfilePicturePath;
+        
+        var user = _useCaseUpdateUserProfilePicture.Execute(dtoInputUpdateProfilePictureUser);
+
+        //Remove the current profile picture if exist
+        if (oldProfilePicturePath != null && oldProfilePicturePath !=
+            "\\Upload\\ProfilePicture\\default_user_pic.png")
+        {
+            _pictureService.RemoveFile(oldProfilePicturePath);
+        }
+
+        return Ok(user);
     }
 
     [HttpPost]
